Factor out real and mock API into container classes inheriting from TopicHighlighter.
