--- app/containers/PybossaHighlighter/styles.scss
@@ -1,21 +1,0 @@
-.topic-picker-wrapper {
-  padding-left: 0 !important;
-}
-
-.article-wrapper {
-  margin: 2em 1em 2em 275px;
-}
-
-.loading {
-  width: 100%;
-  height: 100%;
-  z-index: 100;
-  top: 0px;
-  left: 0px;
-  position: fixed;
-  background-color: rgba(30, 30, 30, 0.3);
-}
-
-.space {
-  height: 100px;
-}
