+++ app/containers/TopicHighlighter/index.js
@@ -1,16 +1,6 @@
-import React, {Component} from 'react';
-import { Link } from 'react-router';
-import { bindActionCreators } from 'redux';
+import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import ReactCSSTransitionsGroup from 'react-addons-css-transition-group';
-import { connect } from 'react-redux';
-
-import * as articleActionCreators from 'actions/article';
-import * as topicsActionCreators from 'actions/topicPicker';
-import * as projectActionCreators from 'actions/project';
-import * as taskActionCreators from 'actions/highlightTasks';
-
-const assembledActionCreators = Object.assign({}, articleActionCreators, topicsActionCreators, projectActionCreators, taskActionCreators)
 
 import { colors } from 'utils/colors';
 import HighlightTool from 'components/HighlightTool';
@@ -18,21 +8,6 @@ import TopicPicker from 'components/TopicPicker';
 import Project from 'components/Project';
 
 import { styles } from './styles.scss';
-
-const mapStateToProps = state => {
-  return {
-    article: state.article.article,
-    saveAndNext: state.article.saveAndNext,
-    currentTopicId: state.topicPicker.currentTopicId,
-    topics: state.topicPicker.topics,
-    highlights: state.highlight.highlights
-  };
-}
-
-@connect (
-  mapStateToProps,
-  dispatch => bindActionCreators(assembledActionCreators, dispatch)
-)
 
 export class TopicHighlighter extends Component {
   constructor(props) {
@@ -44,13 +19,6 @@ export class TopicHighlighter extends Component {
   // will bind 'this' of the class. (React.createClass does automatically.)
   onSaveAndNext = () => {
     this.props.saveAndNext(this.props.highlights);
-  }
-
-  componentDidMount() {
-    this.props.fetchHighlightTasks();
-  }
-
-  componentWillReceiveProps(nextProps) {
   }
 
   render() {
