+++ app/highlight.js
@@ -4,7 +4,7 @@ import React from 'react';
 import { render } from 'react-dom';
 
 import App from 'containers/App';
-import { TopicHighlighter } from 'containers/PybossaHighlighter';
+import { RealHighlighter } from 'containers/HighlighterTasks';
 
 let elem = document.createElement('div');
 elem.id = ('react-root');
@@ -12,7 +12,7 @@ document.body.appendChild(elem);
 
 render(
   <App>
-    <TopicHighlighter/>
+    <RealHighlighter/>
   </App>,
   document.getElementById('react-root')
 );
