+++ app/routes.js
@@ -2,12 +2,12 @@ import React from 'react';
 import { Route } from 'react-router';
 
 import App from 'containers/App';
-import { TopicHighlighter } from 'containers/TopicHighlighter';
+import { MockHighlighter } from 'containers/HighlighterTasks';
 import {Quiz} from 'containers/Quiz';
 
 export default (
   <Route path='/' component={App}>
-    <Route path='article/:articleId' component={TopicHighlighter} />
-    <Route path='quiz/:annotationId' component={Quiz} />
+    <Route path='highlighter' component={MockHighlighter} />
+    <Route path='quiz' component={Quiz} />
   </Route>
 );
